requirements.txt
crewai>=0.60.0
langchain>=0.2.6
openai>=1.40.0
python-dotenv>=1.0.1
duckduckgo-search>=6.2.12
numpy>=2.0.0
pandas>=2.2.2
rich>=13.7.1